requirements.txt
python-dotenv
jsonformatter
python-json-logger
requests
pytz
redis
flask-cors
Flask
Flask-Migrate
Flask-Seeder
sqlalchemy
pyyaml
apispec
marshmallow
greenlet
gevent
serverless_wsgi
boto3
pymysql
elasticsearch>=7.0.0, <8.0
newrelic
influxdb-client
requests==2.29.0 
psycopg2-binary
mysqlclient